Added pre-battle state with card selection
User can cycle through cards with one button and cards are rendered appropriately (although they're overlapping)

Triggers can be cycled with a button, only print to the console.

On sign in user is brought to the card choice screen for the purposes of testing.

WebSocket for battle is opened and data sent on submit but server experiences a type cast error currently.