Added some logic to the battle_system state to read in data from the user.simulation data if it was correctly sent from the server.
Added logic to the pre_battle_state to transition to the battle_system if the meta-server returns proper sim data, otherwise it transitions back to the main menu.

If no response from the meta-server comes after submitting the battle data the player will transition back to the main menu when the battle socket closes.